Fix: Change interventionThreshold type from Float to Int
- Updated TherapistEntity to use @Field(() => Int) instead of Float
- Updated admin resolver argument type to Int
- Fixes GraphQL validation error when Int variables are passed

+++ src/admin/entities/therapist.entity.ts
@@ -1,4 +1,4 @@
-import { Field, ID, ObjectType, Float } from '@nestjs/graphql';
+import { Field, ID, ObjectType, Float, Int } from '@nestjs/graphql';
 
 @ObjectType()
 export class TherapistEntity {
@@ -38,7 +38,7 @@ export class TherapistEntity {
   @Field(() => String, { nullable: true })
   bio?: string | null;
 
-  @Field(() => Float, { nullable: true })
+  @Field(() => Int, { nullable: true })
   interventionThreshold?: number | null;
 
   @Field()
